added application availability check to publish service

+++ newsfeed/src/main/java/newsfeed/ws/UnpublishNewsItemService.java
@@ -10,6 +10,8 @@ import javax.servlet.http.HttpServletResponse;
 import org.apache.log4j.Logger;
 import org.javaimplant.newsfeed.data.NewsItem;
 import org.javaimplant.newsfeed.data.NewsItemDAO;
+import org.javaimplant.newsfeed.data.User;
+import org.javaimplant.newsfeed.data.UserDAO;
 
 public class UnpublishNewsItemService extends HttpServlet {
 	private Logger logger = Logger.getLogger(this.getClass());
@@ -18,8 +20,21 @@ public class UnpublishNewsItemService extends HttpServlet {
 		logger.debug("doGet()");
 		String id = req.getParameter("id");
 		NewsItemDAO newsItemDAO = new NewsItemDAO();
+		// Authenticate client.
+		String accessKey = req.getParameter("accessKey");
+		if (accessKey == null)
+		{
+			resp.sendError(HttpServletResponse.SC_UNAUTHORIZED);
+			return;
+		}
+		User user = new UserDAO().findByAccessKey(accessKey);
+		if(user == null)
+		{
+			resp.sendError(HttpServletResponse.SC_UNAUTHORIZED);
+			return;
+		}
 		NewsItem newsItem = newsItemDAO.find(Long.parseLong(id));
-		if (newsItem != null)
+		if(newsItem != null)
 		{
 			newsItemDAO.delete(newsItem);
 		}
